split out components, fix typo
